requirements.txt
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cryptography==2.8
dialogflow==0.7.2
feedparser==5.2.1
Flask==1.1.1
future==0.18.1
fuzzywuzzy==0.17.0
gnewsclient==1.12
google-api-core==1.14.3
google-auth==1.6.3
googleapis-common-protos==1.6.0
grpcio==1.24.3
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
python-telegram-bot==12.2.0
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.12.0
tornado==6.0.3
urllib3==1.25.6
Werkzeug==0.16.0
